requirements.txt
djangorestframework==3.13.1
pytest-django==4.5.2